+++ lib/src/testing/widget_testing.dart
@@ -2,6 +2,20 @@ import 'package:flutter/material.dart';
 import "package:flutter_test/flutter_test.dart";
 
 
+/// 组件pump
+///
+/// ```dart
+/// Widget widget = ...;
+/// await pumpWidget(tester, widget);
+/// ```
+///
+/// Ref:
+///   - https://docs.flutter.dev/cookbook/testing/widget/introduction#notes-about-the-pump-methods
+Future<dynamic> pumpWidget(WidgetTester tester, Widget widget) async {
+  await tester.pumpWidget(WidgetTestWrapper(widget: widget));
+}
+
+
 /// 异步组件pump
 ///
 /// ```dart